Remove TestFactory
