+++ src/Repository/TestRepository.php
@@ -3,6 +3,8 @@
 namespace App\Repository;
 
 use App\Entity\Test;
+use App\Entity\TestConfiguration;
+use App\Services\EntityStore\TestConfigurationStore;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\ORM\NonUniqueResultException;
 use Doctrine\ORM\NoResultException;
@@ -20,9 +22,31 @@ class TestRepository extends ServiceEntityRepository
 {
     public const DEFAULT_MAX_POSITION = 0;
 
-    public function __construct(ManagerRegistry $registry)
-    {
+    public function __construct(
+        ManagerRegistry $registry,
+        private readonly TestConfigurationStore $configurationStore,
+    ) {
         parent::__construct($registry, Test::class);
+    }
+
+    public function create(
+        TestConfiguration $configuration,
+        string $source,
+        string $target,
+        int $stepCount
+    ): Test {
+        $test = Test::create(
+            $this->configurationStore->get($configuration),
+            $source,
+            $target,
+            $stepCount,
+            $this->findMaxPosition() + 1
+        );
+
+        $this->_em->persist($test);
+        $this->_em->flush();
+
+        return $test;
     }
 
     /**
